refactor: 패키지 생성
