+++ app/src/main/java/woowacourse/movie/fragment/bookhistory/adapter/BookHistoryViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.fragment.bookhistory
+package woowacourse.movie.fragment.bookhistory.adapter
 
 import android.text.SpannableStringBuilder
 import android.text.Spanned
